# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Medigap Plan Price Comparison API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MedigapPlan {
  id: number;
  insurerName: string;
  planLetter: string;
  monthlyPremium: number;
  annualDeductible: number;
  /** @nullable */
  outOfPocketLimit: number | null;
  amBestRating: string;
  /** @nullable */
  moodyRating: string | null;
  yearsInBusiness: number;
  marriedDiscount: boolean;
  /** @nullable */
  notes: string | null;
  planType: string;
  partBDeductibleCovered: boolean;
  foreignTravelCovered: boolean;
}

export interface PlanLetterSummary {
  letter: string;
  count: number;
  minPremium: number;
  maxPremium: number;
  avgPremium: number;
}

export interface PlansSummary {
  state: string;
  county: string;
  totalPlans: number;
  lowestPremium: number;
  highestPremium: number;
  medianPremium: number;
  byLetter: PlanLetterSummary[];
  popularPlanLetter: string;
}

export interface PlanLetterDef {
  letter: string;
  headline: string;
  description: string;
  partACoinsurance: boolean;
  partBCoinsurance: boolean;
  partBDeductible: boolean;
  partADeductible: boolean;
  skilledNursing: boolean;
  foreignTravel: boolean;
  outOfPocketLimit: boolean;
  popularity: string;
}

export interface ZipInfo {
  zip: string;
  city: string;
  state: string;
  stateCode: string;
  county: string;
  region: string;
}

export type GetPlansParams = {
zip: string;
age: number;
married?: boolean;
planLetter?: string;
sortBy?: string;
};

export type GetPlansSummaryParams = {
zip: string;
age: number;
married?: boolean;
};

export type GetZipInfoParams = {
zip: string;
};