# lib/api-zod/src/generated/types/medigapPlan.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Medigap Plan Price Comparison API
 * OpenAPI spec version: 0.1.0
 */

export interface MedigapPlan {
  id: number;
  insurerName: string;
  planLetter: string;
  monthlyPremium: number;
  annualDeductible: number;
  /** @nullable */
  outOfPocketLimit: number | null;
  amBestRating: string;
  /** @nullable */
  moodyRating: string | null;
  yearsInBusiness: number;
  marriedDiscount: boolean;
  /** @nullable */
  notes: string | null;
  planType: string;
  partBDeductibleCovered: boolean;
  foreignTravelCovered: boolean;
}