# lib/api-zod/src/generated/types/planLetterDef.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Medigap Plan Price Comparison API
 * OpenAPI spec version: 0.1.0
 */

export interface PlanLetterDef {
  letter: string;
  headline: string;
  description: string;
  partACoinsurance: boolean;
  partBCoinsurance: boolean;
  partBDeductible: boolean;
  partADeductible: boolean;
  skilledNursing: boolean;
  foreignTravel: boolean;
  outOfPocketLimit: boolean;
  popularity: string;
}