# lib/api-zod/src/generated/types/planLetterSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Medigap Plan Price Comparison API
 * OpenAPI spec version: 0.1.0
 */

export interface PlanLetterSummary {
  letter: string;
  count: number;
  minPremium: number;
  maxPremium: number;
  avgPremium: number;
}