# lib/api-zod/src/generated/types/plansSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Medigap Plan Price Comparison API
 * OpenAPI spec version: 0.1.0
 */
import type { PlanLetterSummary } from './planLetterSummary';

export interface PlansSummary {
  state: string;
  county: string;
  totalPlans: number;
  lowestPremium: number;
  highestPremium: number;
  medianPremium: number;
  byLetter: PlanLetterSummary[];
  popularPlanLetter: string;
}