# lib/api-zod/src/generated/types/zipInfo.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Medigap Plan Price Comparison API
 * OpenAPI spec version: 0.1.0
 */

export interface ZipInfo {
  zip: string;
  city: string;
  state: string;
  stateCode: string;
  county: string;
  region: string;
}